Refactor and enhance PlanMatr API functionality
- Updated `PlanmPartInfo.cs` to use `long?` for `PartId` and added nullable annotations for improved null safety.
- Added `ProductName` and `ShadeName` properties in `PlanogramPartFacing.cs` for better data representation.
- Refined `Include` statements in specifications to optimize data fetching and removed redundant code.
- Expanded `ManagePlanogramsApiController` with new endpoints and improved error handling/logging.
- Enhanced `CreatePlanApiController` and `EditPlanApiController` with new endpoints, async/await usage, and better error handling.
- Reintroduced and expanded mapping profiles with custom resolvers for complex mappings.
- Added new repository services in `PMServiceExtenstions.cs` to support additional functionality.
- Updated logging configuration in `appsettings.json` for better naming consistency.
- Performed extensive refactoring to remove unused code, improve naming conventions, and align with modern practices.
- Introduced new mapping profiles and resolvers for handling entity transformations.
- Improved null safety, logging, and API endpoint consistency across the application.

+++ Dtos/PlanModels/PlanmPartInfo.cs
@@ -36,7 +36,7 @@ namespace PMApplication.Dtos.PlanModels
         //public string ParentPlanxPartId { get; set; }
         public string? PlanxShelfId { get; set; }
 
-        public int? PartId { get; set; }
+        public long? PartId { get; set; }
         public int? PartTypeId { get; set; }
         public int? ShelfTypeId { get; set; }
         public string? PartNumber { get; set; }
